# openapi/panel-api.yaml
openapi: 3.0.3
info:
  title: 3x-ui Panel API
  version: "1.0.0"
  description: |
    OpenAPI specification for the bearer-authenticated panel API exposed under `/panel/api`.

    Notes:
    - All documented endpoints can be called with `Authorization: Bearer <token>`.
    - The implementation also accepts `X-API-Key: <token>` as an alternative.
    - If the panel is served behind a custom `webBasePath`, prepend that base path before `/panel/api`.
    - Most mutation endpoints return HTTP 200 even when `success=false` inside the JSON envelope.
servers:
  - url: "{scheme}://{host}{basePath}"
    description: Panel origin plus optional configured web base path.
    variables:
      scheme:
        default: http
        enum:
          - http
          - https
      host:
        default: localhost:2053
      basePath:
        default: ""
        description: Custom panel base path without trailing slash, for example `/my-panel`.
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /panel/api/server/status:
    get:
      tags: [Server]
      summary: Get server status
      operationId: getServerStatus
      responses:
        "200":
          description: Current server status snapshot.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/Status"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/cpuHistory/{bucket}:
    get:
      tags: [Server]
      summary: Get aggregated CPU history
      operationId: getCpuHistory
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: integer
            enum: [2, 30, 60, 120, 180, 300]
          description: Aggregation bucket size in seconds.
      responses:
        "200":
          description: CPU history points.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: array
                        items:
                          $ref: "#/components/schemas/CPUHistoryPoint"
        "404":
          description: Missing or invalid bearer token.
...
  /panel/api/server/getConfigJson:
    get:
      tags: [Server]
      summary: Get effective Xray configuration JSON
      operationId: getConfigJson
      responses:
        "200":
          description: Free-form JSON configuration object.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/GenericObject"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/getDb:
    get:
      tags: [Server]
      summary: Download the SQLite database
      operationId: downloadDatabase
      responses:
        "200":
          description: SQLite database file.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/getNewUUID:
    get:
      tags: [Server]
      summary: Generate a new UUID
      operationId: getNewUuid
      responses:
        "200":
          description: Generated UUID.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/UuidResult"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/getNewX25519Cert:
    get:
      tags: [Server]
      summary: Generate a new X25519 keypair
      operationId: getNewX25519Cert
      responses:
        "200":
          description: Generated keypair.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/X25519KeyPair"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/getNewmldsa65:
    get:
      tags: [Server]
      summary: Generate a new ML-DSA-65 keypair
      operationId: getNewMldsa65
      responses:
        "200":
          description: Generated keypair.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/Mldsa65KeyPair"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/getNewmlkem768:
    get:
      tags: [Server]
      summary: Generate a new ML-KEM-768 keypair
      operationId: getNewMlkem768
      responses:
        "200":
          description: Generated keypair.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/Mlkem768KeyPair"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/getNewVlessEnc:
    get:
      tags: [Server]
      summary: Generate VLESS encryption presets
      operationId: getNewVlessEnc
      responses:
        "200":
          description: Generated VLESS encryption presets.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/VlessEncResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/stopXrayService:
    post:
      tags: [Server]
      summary: Stop Xray
      operationId: stopXrayService
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/restartXrayService:
    post:
      tags: [Server]
      summary: Restart Xray
      operationId: restartXrayService
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/startVKTurnProxyService:
    post:
      tags: [Server]
      summary: Start all enabled vk-turn-proxy processes
      operationId: startVkTurnProxyService
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/stopVKTurnProxyService:
    post:
      tags: [Server]
      summary: Stop all vk-turn-proxy processes and pause auto-restart
      operationId: stopVkTurnProxyService
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/restartVKTurnProxyService:
    post:
      tags: [Server]
      summary: Restart all enabled vk-turn-proxy processes
      operationId: restartVkTurnProxyService
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/installXray/{version}:
    post:
      tags: [Server]
      summary: Install or switch to a specific Xray version
      operationId: installXray
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/updateGeofile:
    post:
      tags: [Server]
      summary: Update all supported geofiles
      operationId: updateGeofile
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/updateGeofile/{fileName}:
    post:
      tags: [Server]
      summary: Update one specific geofile
      operationId: updateSingleGeofile
      parameters:
        - name: fileName
          in: path
          required: true
          schema:
            type: string
          description: Filename validated by the server against a safe character whitelist.
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/logs/{count}:
    post:
      tags: [Server]
      summary: Get panel logs
      operationId: getLogs
      parameters:
        - name: count
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/LogsRequest"
      responses:
        "200":
          description: Log lines.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: array
                        items:
                          type: string
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/xraylogs/{count}:
    post:
      tags: [Server]
      summary: Get parsed Xray access logs
      operationId: getXrayLogs
      parameters:
        - name: count
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/XrayLogsRequest"
      responses:
        "200":
          description: Parsed Xray log entries.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: array
                        items:
                          $ref: "#/components/schemas/LogEntry"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/vkturnproxylogs/{count}:
    post:
      tags: [Server]
      summary: Get vk-turn-proxy process logs from the panel log buffer
      operationId: getVkTurnProxyLogs
      parameters:
        - name: count
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                level:
                  type: string
                  enum: [debug, info, notice, warning, err]
      responses:
        "200":
          description: Filtered vk-turn-proxy log lines.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: array
                        items:
                          type: string
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/importDB:
    post:
      tags: [Server]
      summary: Import a database backup
      operationId: importDatabase
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [db]
              properties:
                db:
                  type: string
                  format: binary
      responses:
        "200":
          description: "Import result. `obj` contains a localized success string."
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: string
        "404":
          description: Missing or invalid bearer token.
  /panel/api/server/getNewEchCert:
    post:
      tags: [Server]
      summary: Generate an ECH certificate set
      operationId: getNewEchCert
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/EchCertRequest"
      responses:
        "200":
          description: Generated ECH values.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/EchCertResult"
        "404":
          description: Missing or invalid bearer token.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
      description: "Pass the panel API token as `Authorization: Bearer <token>`."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alternative header accepted by the implementation.
...
  responses:
    ApiEnvelopeResponse:
      description: Standard JSON envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
  schemas:
    ApiEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        msg:
          type: string
        obj:
          nullable: true
    GenericObject:
      type: object
      additionalProperties: true
...
    LogsRequest:
      type: object
      properties:
        level:
          type: string
        syslog:
          type: string
          description: Use `"true"` to read journalctl output instead of application logs.
    XrayLogsRequest:
      type: object
      properties:
        filter:
          type: string
        showDirect:
          type: string
          description: String boolean as used by the handler, for example `"true"` or `"false"`.
        showBlocked:
          type: string
        showProxy:
          type: string
    EchCertRequest:
      type: object
      required: [sni]
      properties:
        sni:
          type: string
...
    ProcessStatusSummary:
      type: object
      properties:
        available:
          type: boolean
        total:
          type: integer
        enabled:
          type: integer
        running:
          type: integer
        state:
          type: string
          enum: [running, stop, error]
        errorMsg:
          type: string
        uptime:
          type: integer
          format: int64
    Status:
      type: object
      properties:
        cpu:
          type: number
          format: double
        cpuCores:
          type: integer
        logicalPro:
          type: integer
        cpuSpeedMhz:
          type: number
          format: double
        mem:
          $ref: "#/components/schemas/CurrentTotalStats"
        swap:
          $ref: "#/components/schemas/CurrentTotalStats"
        disk:
          $ref: "#/components/schemas/CurrentTotalStats"
        xray:
          type: object
          properties:
            state:
              type: string
              enum: [running, stop, error]
            errorMsg:
              type: string
            version:
              type: string
        vkTurnProxy:
          $ref: "#/components/schemas/ProcessStatusSummary"
        uptime:
          type: integer
          format: int64
        loads:
          type: array
          items:
            type: number
            format: double
        tcpCount:
          type: integer
        udpCount:
          type: integer
        netIO:
          type: object
          properties:
            up:
              type: integer
              format: int64
            down:
              type: integer
              format: int64
        netTraffic:
          type: object
          properties:
            sent:
              type: integer
              format: int64
            recv:
              type: integer
              format: int64
        publicIP:
          type: object
          properties:
            ipv4:
              type: string
            ipv6:
              type: string
        appStats:
          type: object
          properties:
            threads:
              type: integer
            mem:
              type: integer
              format: int64
            uptime:
              type: integer
              format: int64
    CurrentTotalStats:
      type: object
      properties:
        current:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
    CPUHistoryPoint:
      type: object
      properties:
        t:
          type: integer
          format: int64
        cpu:
          type: number
          format: double
    LogEntry:
      type: object
      properties:
        DateTime:
          type: string
          format: date-time
        FromAddress:
          type: string
        ToAddress:
          type: string
        Inbound:
          type: string
        Outbound:
          type: string
        Email:
          type: string
        Event:
          type: integer
          description: 0=direct, 1=blocked, 2=proxied
    UuidResult:
      type: object
      properties:
        uuid:
          type: string
    X25519KeyPair:
      type: object
      properties:
        privateKey:
          type: string
        publicKey:
          type: string
    Mldsa65KeyPair:
      type: object
      properties:
        seed:
          type: string
        verify:
          type: string
    Mlkem768KeyPair:
      type: object
      properties:
        seed:
          type: string
        client:
          type: string
    EchCertResult:
      type: object
      properties:
        echServerKeys:
          type: string
        echConfigList:
          type: string
    VlessEncResponse:
      type: object
      properties:
        auths:
          type: array
          items:
            $ref: "#/components/schemas/VlessEncAuth"
    VlessEncAuth:
      type: object
      properties:
        label:
          type: string
        decryption:
          type: string
        encryption:
          type: string